+++ src/utils.ts
@@ -981,11 +981,7 @@ export async function processTransactionBloxroute(
 
   let failures = 0;
   while (true) {
-    let recentBlockhash =
-      blockhash ??
-      (await anchorProvider.connection.getLatestBlockhash(
-        Exchange.blockhashCommitment
-      ));
+    let recentBlockhash = blockhash ?? (await Exchange.getCachedBlockhash());
 
     let v0Tx: VersionedTransaction = new VersionedTransaction(
       new TransactionMessage({
@@ -1108,11 +1104,17 @@ export async function processTransactionJito(
     );
   }
 
-  let recentBlockhash =
-    blockhash ??
-    (await provider.connection.getLatestBlockhash(
-      Exchange.blockhashCommitment
-    ));
+  tx.instructions.push(
+    SystemProgram.transfer({
+      fromPubkey: Exchange.provider.publicKey,
+      toPubkey: new PublicKey(
+        "DttWaMuVvTiduZRnguLF7jNxTgiMBZ1hyAumKUiL2KRL" // Jito tip account
+      ),
+      lamports: Exchange.jitoTip, // tip
+    })
+  );
+
+  let recentBlockhash = blockhash ?? (await Exchange.getCachedBlockhash());
 
   let vTx: VersionedTransaction = new VersionedTransaction(
     new TransactionMessage({
@@ -1203,6 +1205,11 @@ export async function processTransactionJito(
 
   while (currentBlockHeight < recentBlockhash.lastValidBlockHeight) {
     // Keep resending to maximise the chance of confirmation
+    await provider.connection.sendRawTransaction(rawTx, {
+      skipPreflight: true,
+      preflightCommitment: provider.connection.commitment,
+      maxRetries: 0,
+    });
 
     if (comboRpc) {
       await provider.connection.sendRawTransaction(rawTx, {
@@ -1267,6 +1274,34 @@ export async function processTransactionJito(
   }
 }
 
+export async function sendRawTransactionCaught(con: Connection, rawTx: any) {
+  try {
+    let txSig = await con.sendRawTransaction(rawTx, {
+      skipPreflight: true,
+      preflightCommitment: con.commitment,
+      maxRetries: 0,
+    });
+    return txSig;
+  } catch (e) {
+    console.log(`Error sending tx: ${e}`);
+  }
+}
+
+export async function sendJitoTxCaught(payload: any) {
+  try {
+    let response = await axios.post(
+      "https://mainnet.block-engine.jito.wtf/api/v1/transactions",
+      payload,
+      {
+        headers: { "Content-Type": "application/json" },
+      }
+    );
+    return response.data.result;
+  } catch (e) {
+    console.log(`Error sending tx: ${e}`);
+  }
+}
+
 export async function processTransaction(
   provider: anchor.AnchorProvider,
   tx: Transaction,
@@ -1318,11 +1353,7 @@ export async function processTransaction(
       );
     }
 
-    let recentBlockhash =
-      blockhash ??
-      (await provider.connection.getLatestBlockhash(
-        Exchange.blockhashCommitment
-      ));
+    let recentBlockhash = blockhash ?? (await Exchange.getCachedBlockhash());
 
     if (lutAccs) {
       if (useLedger) {
@@ -1369,6 +1400,9 @@ export async function processTransaction(
 
     let txOpts = opts || commitmentConfig(provider.connection.commitment);
     let txSig;
+    let allConnections = [provider.connection].concat(
+      Exchange.doubleDownConnections
+    );
     try {
       // Integration tests don't like the split send + confirm :(
       if (Exchange.network == Network.LOCALNET) {
@@ -1379,21 +1413,44 @@ export async function processTransaction(
         );
       }
 
-      txSig = await provider.connection.sendRawTransaction(rawTx, {
-        skipPreflight: true,
-        preflightCommitment: provider.connection.commitment,
-      });
+      let promises = [];
+      for (var con of allConnections) {
+        promises.push(sendRawTransactionCaught(con, rawTx));
+      }
+
+      // Jito's transactions endpoint, not a bundle
+      // Might as well send it here for extra success, it's free
+      if (Exchange.network == Network.MAINNET) {
+        const encodedTx = bs58.encode(rawTx);
+        const payload = {
+          jsonrpc: "2.0",
+          id: 1,
+          method: "sendTransaction",
+          params: [encodedTx],
+        };
+        promises.push(sendJitoTxCaught(payload));
+      }
+
+      // All tx sigs are the same
+      let txSigs = await Promise.all(promises);
+      let txSig = txSigs.find(
+        (sig) => typeof sig === "string" && sig.length > 0
+      );
 
       // Poll the tx confirmation for N seconds
       // Polling is more reliable than websockets using confirmTransaction()
       let currentBlockHeight = 0;
       if (!Exchange.skipRpcConfirmation) {
+        let resendCounter = 0;
         while (currentBlockHeight < recentBlockhash.lastValidBlockHeight) {
           // Keep resending to maximise the chance of confirmation
-          await provider.connection.sendRawTransaction(rawTx, {
-            skipPreflight: true,
-            preflightCommitment: provider.connection.commitment,
-          });
+          resendCounter += 1;
+          if (resendCounter % 4 == 0) {
+            for (var con of allConnections) {
+              promises.push(sendRawTransactionCaught(con, rawTx));
+            }
+            await Promise.race(promises);
+          }
 
           let status = await provider.connection.getSignatureStatus(txSig);
           currentBlockHeight = await provider.connection.getBlockHeight(
@@ -2564,7 +2621,7 @@ async function initializeZetaMarket(
   const bids = getOrCreateKeypair(`${dir}/bids-${i}.json`);
   const asks = getOrCreateKeypair(`${dir}/asks-${i}.json`);
 
-  let [preTx, tx, tx2, postTx] = await instructions.initializeZetaMarketTxs(
+  let [tx, tx2] = await instructions.initializeZetaMarketTxs(
     asset,
     marketIndexesAccount.indexes[i],
     requestQueue.publicKey,
@@ -2601,16 +2658,7 @@ async function initializeZetaMarket(
     console.log(`Market ${i} serum accounts already initialized...`);
   } else {
     try {
-      console.log("initialize market pda");
-      await processTransaction(
-        Exchange.provider,
-        preTx,
-        [],
-        commitmentConfig(Exchange.connection.commitment),
-        Exchange.useLedger
-      );
-
-      console.log("initialize zeta market instruction");
+      console.log("initialize zeta market accounts");
       await processTransaction(
         Exchange.provider,
         tx,
@@ -2627,17 +2675,10 @@ async function initializeZetaMarket(
     console.log(`Market ${i} already initialized. Skipping...`);
   } else {
     try {
+      console.log("initialize zeta market instruction");
       await processTransaction(
         Exchange.provider,
         tx2,
-        [],
-        commitmentConfig(Exchange.connection.commitment),
-        Exchange.useLedger
-      );
-
-      await processTransaction(
-        Exchange.provider,
-        postTx,
         [],
         commitmentConfig(Exchange.connection.commitment),
         Exchange.useLedger